// apps/web/src/app/features/inventory/modules/drug-focus-lists/types/drug-focus-lists.types.ts
export interface DrugFocusList {
  id: number;
  list_code: string;
  list_name: string;
  description?: string | null;
  generic_id?: number | null;
  drug_id?: number | null;
  is_active?: boolean | null;
  created_at?: string | null;
  updated_at?: string | null;
}